+++ crates/air/src/commands/format.rs
@@ -1,4 +1,9 @@
+use std::path::Path;
+
 use workspace::discovery;
+use workspace::discovery::DiscoveredSettings;
+use workspace::resolve::PathResolver;
+use workspace::settings::Settings;
 
 use crate::ExitStatus;
 use crate::args::FormatCommand;
@@ -12,12 +17,19 @@ enum FormatMode {
     Check,
 }
 
+#[derive(Copy, Clone, Debug)]
+enum ConfigurationMode {
+    Discover,
+    Disabled,
+}
+
 pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
     if let Some(status) = check_argument_consistency(&command) {
         return Ok(status);
     }
 
     let mode = FormatMode::from_command(&command);
+    let configuration = ConfigurationMode::from_command(&command);
 
     let (exclude, include) = if command.force {
         (discovery::Exclude::Nothing, discovery::Include::Everything)
@@ -26,9 +38,31 @@ pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
     };
 
     match command.stdin_file_path {
-        Some(path) => stdin::format(path, mode, exclude, include),
-        None => paths::format(command.paths, mode, exclude, include),
+        Some(path) => stdin::format(path, mode, configuration, exclude, include),
+        None => paths::format(command.paths, mode, configuration, exclude, include),
     }
+}
+
+fn resolve_settings<P: AsRef<Path>>(
+    paths: &[P],
+    configuration: ConfigurationMode,
+) -> anyhow::Result<(PathResolver<Settings>, Settings)> {
+    let mut resolver = PathResolver::new();
+
+    if let ConfigurationMode::Discover = configuration {
+        for DiscoveredSettings {
+            directory,
+            settings,
+        } in discovery::discover_settings(paths)?
+        {
+            resolver.add(&directory, settings);
+        }
+
+        let default_settings = discovery::discover_user_settings()?.unwrap_or_default();
+        return Ok((resolver, default_settings));
+    }
+
+    Ok((resolver, Settings::default()))
 }
 
 fn check_argument_consistency(command: &FormatCommand) -> Option<ExitStatus> {
@@ -54,6 +88,16 @@ impl FormatMode {
             FormatMode::Check
         } else {
             FormatMode::Write
+        }
+    }
+}
+
+impl ConfigurationMode {
+    fn from_command(command: &FormatCommand) -> Self {
+        if command.no_configuration {
+            ConfigurationMode::Disabled
+        } else {
+            ConfigurationMode::Discover
         }
     }
 }