+++ crates/air/src/commands/format/stdin.rs
@@ -8,9 +8,6 @@ use std::path::PathBuf;
 
 use thiserror::Error;
 use workspace::discovery;
-use workspace::discovery::DiscoveredSettings;
-use workspace::discovery::discover_settings;
-use workspace::discovery::discover_user_settings;
 use workspace::format::FormatSourceError;
 use workspace::format::FormattedSource;
 use workspace::resolve::PathResolver;
@@ -18,7 +15,9 @@ use workspace::settings::FormatSettings;
 use workspace::settings::Settings;
 
 use crate::ExitStatus;
+use crate::commands::format::ConfigurationMode;
 use crate::commands::format::FormatMode;
+use crate::commands::format::resolve_settings;
 
 #[derive(Debug)]
 enum FormattedStdin {
@@ -38,23 +37,14 @@ enum FormatStdinError {
 pub(crate) fn format(
     path: PathBuf,
     mode: FormatMode,
+    configuration: ConfigurationMode,
     exclude: discovery::Exclude,
     include: discovery::Include,
 ) -> anyhow::Result<ExitStatus> {
     // Normalize up front, relative to current working directory
     let path = fs::normalize_path(path);
 
-    let mut resolver = PathResolver::new();
-
-    for DiscoveredSettings {
-        directory,
-        settings,
-    } in discover_settings(&[&path])?
-    {
-        resolver.add(&directory, settings);
-    }
-
-    let default_settings = discover_user_settings()?.unwrap_or_default();
+    let (resolver, default_settings) = resolve_settings(&[&path], configuration)?;
 
     match mode {
         FormatMode::Write => {